requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
pydantic==1.10.4
python-multipart==0.0.5
python-dotenv==0.21.0
SQLAlchemy==1.4.46
alembic==1.9.2
PyMySQL==1.0.2
PyJWT==2.6.0
